+++ src/Resources/contao/dca/tl_wertungsportal_besucher.php
@@ -13,7 +13,8 @@ $GLOBALS['TL_DCA']['tl_wertungsportal_besucher'] = array
 (
 	'config' => array
 	(
-		'dataContainer'    => 'Table',
+		// Klassenname statt der Kurzform 'Table': Contao 5 kennt sie nicht mehr
+		'dataContainer'    => \Contao\DC_Table::class,
 		'closed'           => true,
 		'notEditable'      => true,
 		'sql' => array